refactor: move all query and mutations strings to constants
