+++ packages/api/test/utils/constants.ts
@@ -1,0 +1,75 @@
+export const statusQuery = `
+        query {
+            status
+        }
+        `;
+
+export const usersQuery = `
+query {
+    users {
+        id
+        firstName
+        lastName
+        displayName
+        email
+    }
+}`;
+
+export const userQuery = `
+query getUserById($id: String!) {
+    user(id: $id) {
+        id
+        firstName
+        lastName
+        displayName
+        email
+    }
+}`;
+
+export const RegisterUserOp = `
+mutation RegisterUser($userData: RegisterUserInput!) {
+    registerUser(userData: $userData) {
+        id
+        firstName
+        lastName
+        displayName
+        email
+    }
+}`;
+
+export const meQuery = `
+query {
+    me {
+        id
+        firstName
+        lastName
+        displayName
+        email
+    }
+}`;
+
+export const loginOp = `
+mutation Login($email: String!, $password: String!) {
+    login(email: $email, password: $password) {
+        accessToken
+    }
+}`;
+
+export const logoutOp = `
+mutation Logout {
+    logout
+}`;
+
+export const refreshAccessTokenOp = `
+mutation RefreshAccessToken {
+    refreshAccessToken {
+        accessToken
+    }
+}
+`;
+
+export const avatarUploadOp = `
+mutation AvatarUpload ($image: Upload!) {
+    avatarUpload(image: $image)
+}
+`;
